// mobile-app/src/types/entities.ts
export interface Transactions {
    ID: string;
    account?: Accounts;
    account_ID?: string;
    currency: string;
    amount: number;
    date: Date;
    description: string;
    mcc?: MCC;
    mcc_ID?: string;
    CO2Score?: number;
}

export interface AccountHabits {
    ID: string;
    account?: Accounts;
    account_ID?: string;
    habits?: Habits;
    habits_ID?: string;
    transaction?: Transactions;
    transaction_ID?: string;
}

export interface Accounts {
    ID: string;
    name: string;
    transactions: Transactions[];
    challenges: ChallengesUsers[];
    habits: AccountHabits[];
}
// ...
export interface Challenges {
    ID: string;
    title: string;
    description: string;
    mcc?: MCC;
    mcc_ID?: string;
    daysToMark: number;
    avoidableEmissionsPerDay: number;
}

export interface ChallengesUsers {
    ID: string;
    account?: Accounts;
    account_ID?: string;
    challenge?: Challenges;
    challenge_ID?: string;
    startDate: Date;
    dueDate: Date;
    markedDates: Date[];
    isCompleted: boolean;
}

export interface Categories {
    ID: string;
    description: string;
    icon: string;
    mccs: MCC[];
}

export interface MCC {
    ID: string;
    MCC: string;
    description: string;
    factor: number;
    category?: Categories;
    category_ID?: string;
}